Master ng

+++ src/main/scala/com/wavesplatform/settings/FeaturesSettings.scala
@@ -2,4 +2,4 @@ package com.wavesplatform.settings
 
 case class FeaturesSettings(autoActivate: Boolean,
                             autoShutdownOnUnsupportedFeature: Boolean,
-                            supported: List[Int])
+                            supported: List[Short])
